main/module/requirement.txt
pandas 
numpy
re
matplotlib.pyplot
seaborn
sklearn
